+++ src/tasks/yq.rs
@@ -6,13 +6,13 @@ use utils::github::{Asset, Release};
 use utils::golang::{arch, os};
 
 pub fn sync() {
-    println!("pkg: shfmt: syncing ...");
+    println!("pkg: yq: syncing ...");
 
     if !is_installed() {
-        let release = match utils::github::latest_release(&"mvdan", &"sh") {
+        let release = match utils::github::latest_release(&"mikefarah", &"yq") {
             Ok(r) => r,
             Err(error) => {
-                println!("error: pkg: shfmt: {}", error);
+                println!("error: pkg: yq: {}", error);
                 return;
             }
         };
@@ -25,13 +25,13 @@ pub fn update() {
         return;
     }
 
-    println!("pkg: shfmt: updating ...");
+    println!("pkg: yq: updating ...");
 
-    match utils::process::command_output("shfmt", &["--version"]) {
+    match utils::process::command_output("yq", &["--version"]) {
         Ok(output) => {
             let stdout = std::str::from_utf8(&output.stdout).unwrap_or_default();
 
-            match utils::github::release_versus_current(&stdout, &"mvdan", &"sh") {
+            match utils::github::release_versus_current(&stdout, &"mikefarah", &"yq") {
                 Some(r) => install_release_asset(r),
                 None => {}
             }
@@ -44,24 +44,23 @@ fn install_release_asset(release: Release) {
     let asset = match latest_asset(&release) {
         Some(a) => a,
         None => {
-            println!("pkg: shfmt: no asset matches OS and ARCH");
+            println!("pkg: yq: no asset matches OS and ARCH");
             return;
         }
     };
 
-    println!("pkg: shfmt: installing ...");
+    println!("pkg: yq: installing ...");
 
-    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/shfmt"));
+    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/yq"));
     utils::github::download_release_asset(asset, &bin_path);
 }
 
 fn is_installed() -> bool {
-    match utils::process::command_output("shfmt", &["--version"]) {
+    match utils::process::command_output("yq", &["--version"]) {
         Ok(output) => output.status.success(),
         Err(_error) => false,
     }
 }
-
 
 fn latest_asset(release: &Release) -> Option<Asset> {
     return release
@@ -70,9 +69,9 @@ fn latest_asset(release: &Release) -> Option<Asset> {
         .into_iter()
         .filter_map(|asset| {
             #[cfg(windows)]
-            let name = format!("shfmt_{}_{}_{}.exe", release.tag_name, os(), arch());
+            let name = format!("yq_{}_{}.exe", os(), arch());
             #[cfg(not(windows))]
-            let name = format!("shfmt_{}_{}_{}", release.tag_name, os(), arch());
+            let name = format!("yq_{}_{}", os(), arch());
 
             if asset.name == name {
                 Some(asset)