+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* `utils::http::fetch_request()` follows redirects
+
 ## [0.8.0] - 2018-04-17
 
 ### Added